requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
openai==1.57.4
packaging==24.2
pluggy==1.5.0
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
sniffio==1.3.1
todoist_api_python==2.1.7
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
